add maven bot (#27)
and switch to java.time (#28)

+++ src/test/java/net/aclrian/mpe/controller/TestFinishController.java
@@ -1,42 +1,28 @@
 package net.aclrian.mpe.controller;
 
-import javafx.application.Platform;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Node;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.ListView;
-import javafx.scene.layout.Pane;
-import javafx.stage.Stage;
-import javafx.stage.Window;
-import javafx.util.Pair;
-import net.aclrian.mpe.messdiener.Messdaten;
-import net.aclrian.mpe.messdiener.Messdiener;
-import net.aclrian.mpe.messe.Messe;
-import net.aclrian.mpe.messe.Messverhalten;
-import net.aclrian.mpe.messe.StandartMesse;
-import net.aclrian.mpe.pfarrei.Einstellungen;
-import net.aclrian.mpe.pfarrei.Pfarrei;
-import net.aclrian.mpe.utils.DateienVerwalter;
-import net.aclrian.mpe.utils.Dialogs;
-import net.aclrian.mpe.utils.Log;
-import org.junit.Test;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.testfx.assertions.api.Assertions;
-import org.testfx.framework.junit.ApplicationTest;
-import org.testfx.util.WaitForAsyncUtils;
+import javafx.application.*;
+import javafx.fxml.*;
+import javafx.scene.*;
+import javafx.scene.control.*;
+import javafx.scene.layout.*;
+import javafx.stage.*;
+import javafx.util.*;
+import net.aclrian.mpe.messdiener.*;
+import net.aclrian.mpe.messe.*;
+import net.aclrian.mpe.pfarrei.*;
+import net.aclrian.mpe.utils.*;
+import org.junit.*;
+import org.mockito.*;
+import org.testfx.assertions.api.*;
+import org.testfx.framework.junit.*;
+import org.testfx.util.*;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.nio.file.Files;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
+import java.io.*;
+import java.net.*;
+import java.nio.file.*;
+import java.time.*;
+import java.time.format.*;
+import java.util.*;
 
 public class TestFinishController extends ApplicationTest {
 
@@ -121,12 +107,12 @@ public class TestFinishController extends ApplicationTest {
                 instance.afterstartup(pane.getScene().getWindow(), mc);
             } catch (IOException e) {
                 Log.getLogger().error(e.getMessage(), e);
-                Assertions.fail("Could not open " + MainController.EnumPane.FERIEN.getLocation() + e.getLocalizedMessage());
+                Assertions.fail("Could not open " + MainController.EnumPane.FERIEN.getLocation() + e.getLocalizedMessage(), e);
             }
         });
         WaitForAsyncUtils.waitForFxEvents();
 
-        Date date = TestFerienplanController.getToday();
+        LocalDate date = TestFerienplanController.getToday();
         m1.getMessdaten().einteilenVorzeitig(date, false);
 
 
@@ -134,16 +120,14 @@ public class TestFinishController extends ApplicationTest {
                 .thenReturn(Dialogs.YesNoCancelEnum.CANCEL, Dialogs.YesNoCancelEnum.YES));
         WaitForAsyncUtils.waitForFxEvents();
         Platform.runLater(() -> {
-            if (pane.lookup("#zurueck") instanceof Button) {
-                Button zurueck = (Button) pane.lookup("#zurueck");
+            if (pane.lookup("#zurueck") instanceof Button zurueck) {
                 zurueck.fire();
             }
         });
         WaitForAsyncUtils.waitForFxEvents();
         Assertions.assertThat(m1.getMessdaten().kanndann(date, false)).isFalse();
         Platform.runLater(() -> {
-            if (pane.lookup("#zurueck") instanceof Button) {
-                Button zurueck = (Button) pane.lookup("#zurueck");
+            if (pane.lookup("#zurueck") instanceof Button zurueck) {
                 zurueck.fire();
             } else {
                 Assertions.fail("Could not find zurueck");
@@ -161,8 +145,7 @@ public class TestFinishController extends ApplicationTest {
         Assertions.assertThat(instance.getNichtEingeteile()).contains(m3);
 
         Platform.runLater(() -> {
-            if (pane.lookup("#nichteingeteilte") instanceof Button) {
-                Button nichteingeteilte = (Button) pane.lookup("#nichteingeteilte");
+            if (pane.lookup("#nichteingeteilte") instanceof Button nichteingeteilte) {
                 nichteingeteilte.fire();
             } else {
                 Assertions.fail("Could not find nichteingeteilte");
@@ -185,13 +168,22 @@ public class TestFinishController extends ApplicationTest {
         WaitForAsyncUtils.waitForFxEvents();
     }
 
-    //Ignore //turn it back
     @Test
-    public void testEinteilung() {
+    public void testEinteilungWithoutDOCX() {
+        testEinteilung(false);
+    }
+
+    @Ignore("Run Test with conversion to docx")
+    @Test
+    public void testEinteilungWithDOCX() {
+        testEinteilung(true);
+    }
+
+    private void testEinteilung(boolean skipDOCX) {
         Messdiener m1Freund = Mockito.mock(Messdiener.class);
         Mockito.when(dv.getMessdiener()).thenReturn(Arrays.asList(m1, m2, m3, m1Freund));
         Mockito.when(dv.getPfarrei()).thenReturn(pf);
-        StandartMesse standartMesse = new StandartMesse("Mo", 8, "00", "o1", 2, "t1");
+        StandartMesse standartMesse = new StandartMesse(DayOfWeek.MONDAY, 8, "00", "o1", 2, "t1");
         Mockito.when(pf.getStandardMessen()).thenReturn(Collections.singletonList(standartMesse));
         Mockito.when(pf.getSettings()).thenReturn(einst);
         einst.editMaxDienen(false, 10);
@@ -250,9 +242,9 @@ public class TestFinishController extends ApplicationTest {
         Mockito.when(m3.getDienverhalten()).thenReturn(mv3);
         Mockito.when(m1Freund.getDienverhalten()).thenReturn(mv1F);
 
-        Messe me1 = new Messe(false, 1, TestFerienplanController.getYesterday2(), "o1", "t1");
-        Messe me2 = new Messe(TestFerienplanController.getToday(), standartMesse);
-        Messe me3 = new Messe(false, 1, TestFerienplanController.getTomorrow(), "o3", "t3");
+        Messe me1 = new Messe(false, 1, TestFerienplanController.getYesterday2().atTime(0, 0), "o1", "t1");
+        Messe me2 = new Messe(TestFerienplanController.getToday().atTime(0, 0), standartMesse);
+        Messe me3 = new Messe(false, 1, TestFerienplanController.getTomorrow().atTime(0, 0), "o3", "t3");
 
         Mockito.doCallRealMethod().when(dialogs).show(Mockito.anyList(), Mockito.eq(FinishController.NICHT_EINGETEILTE_MESSDIENER));
 
@@ -271,7 +263,7 @@ public class TestFinishController extends ApplicationTest {
                 instance.afterstartup(pane.getScene().getWindow(), mc);
             } catch (IOException e) {
                 Log.getLogger().error(e.getMessage(), e);
-                Assertions.fail("Could not open " + MainController.EnumPane.FERIEN.getLocation() + e.getLocalizedMessage());
+                Assertions.fail("Could not open " + MainController.EnumPane.FERIEN.getLocation() + e.getLocalizedMessage(), e);
             }
         });
         WaitForAsyncUtils.waitForFxEvents();
@@ -292,27 +284,28 @@ public class TestFinishController extends ApplicationTest {
             Files.delete(out.toPath());
         } catch (IOException e) {
             Log.getLogger().error(e.getMessage(), e);
-            Assertions.fail(e.getMessage());
+            Assertions.fail(e.getMessage(), e);
         }
-        out = new File(Log.getWorkingDir(), instance.getTitle() + ".docx");
-        try {
-            if (out.exists()) {
+        if (skipDOCX) {
+            out = new File(Log.getWorkingDir(), instance.getTitle() + ".docx");
+            try {
+                if (out.exists()) {
+                    Files.delete(out.toPath());
+                }
+                instance.toWORD(null);
+                Assertions.assertThat(out.exists()).isTrue();
                 Files.delete(out.toPath());
+            } catch (IOException e) {
+                Log.getLogger().error(e.getMessage(), e);
+                Assertions.fail(e.getMessage(), e);
             }
-            instance.toWORD(null);
-            Assertions.assertThat(out.exists()).isTrue();
-            Files.delete(out.toPath());
-        } catch (IOException e) {
-            Log.getLogger().error(e.getMessage(), e);
-            Assertions.fail(e.getMessage());
         }
 
         Platform.runLater(() -> Mockito.when(dialogs.yesNoCancel(Mockito.anyString(), Mockito.anyString(), Mockito.anyString(), Mockito.anyString()))
                 .thenReturn(Dialogs.YesNoCancelEnum.YES));
         WaitForAsyncUtils.waitForFxEvents();
         Platform.runLater(() -> {
-            if (pane.lookup("#zurueck") instanceof Button) {
-                Button zurueck = (Button) pane.lookup("#zurueck");
+            if (pane.lookup("#zurueck") instanceof Button zurueck) {
                 zurueck.fire();
             } else {
                 Assertions.fail("Could not find zurueck");
@@ -328,8 +321,12 @@ public class TestFinishController extends ApplicationTest {
         Assertions.assertThat(me2.getEingeteilte().size()).isEqualTo(0);
         Assertions.assertThat(me3.getEingeteilte().size()).isEqualTo(0);
 
+        int from = TestFerienplanController.getYesterday2().getDayOfMonth();
+        int to = TestFerienplanController.getTomorrow().getDayOfMonth();
+        String fromMonth = TestFerienplanController.getYesterday2().getMonth().getDisplayName(TextStyle.FULL, Locale.getDefault());
+        String toMonth = TestFerienplanController.getTomorrow().getMonth().getDisplayName(TextStyle.FULL, Locale.getDefault());
         Pair<List<Messdiener>, StringBuilder> pair = instance.getResourcesForEmail();
-        Assertions.assertThat(pair.getValue()).contains("mailto:");
+        Assertions.assertThat(pair.getValue().toString()).isEqualTo("mailto:?bcc=a@w.de&subject=Messdienerplan%20vom%20" + from + ".%20" + fromMonth + "%20bis%20" + to + ".%20" + toMonth + "&body=%0D%0A");
         try {
             new URI(pair.getValue().toString());
         } catch (URISyntaxException e) {
@@ -337,14 +334,5 @@ public class TestFinishController extends ApplicationTest {
         }
 
         Assertions.assertThat(pair.getKey()).hasSize(3).containsAll(Arrays.asList(m1, m2, m1Freund));
-
-        Platform.runLater(() -> {
-            if (pane.lookup("#mail") instanceof Button) {
-                Button mail = (Button) pane.lookup("#mail");
-                mail.fire();
-            } else {
-                Assertions.fail("Could not find mail");
-            }
-        });
     }
 }
